+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - concord-server, concord-console: support for `yield` for
 non-custom forms;
+- ansible: support for external private key files;
 - ansible: support for external configuration files;
 - ansible: verbosity level can now be set using the task's arguments.